Storage as a Set.
Turns out you do need our storage to be a Set instead of a list so that it can easily distinguish between duplicates. Then, when we want to sort it, we can stream it into a list.

Also, I defined the order I want the XML elements marshalled in.

Also, I've got the run configuration stored as a file for ease of running.